requirements.txt
# Wechaty and related packages
wechaty==0.10.7
wechaty-puppet-padlocal>=0.4.1
python-wechaty-plugin-contrib>=0.0.8

# Web framework
flask==2.0.1
flask-cors==3.0.10

# Database
sqlalchemy==1.4.27
alembic==1.7.5

# AI and NLP
openai==0.27.0
jieba==0.42.1

# Data processing
pandas==1.3.4
numpy==1.21.4

# Visualization
matplotlib==3.5.0
plotly==5.5.0

# Utilities
python-dotenv==0.19.2
requests==2.26.0
pydantic==1.8.2
python-dateutil==2.8.2
pytz==2021.3
schedule==1.1.0
qrcode==7.3.1
pillow==8.4.0

# Development tools
pytest==6.2.5
black==21.10b0
flake8==4.0.1 